Add support for both v13 and v14 of Foundry
